# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: AsterFerry Controller API
  version: 1.0.0
  description: Authoritative control-plane API for the AsterFerry v1.0.0 normalized architecture. Gateway and Agent business traffic never traverses this API.
servers:
  - url: /api/v1
...
paths:
...
  /nodes:
    get:
      tags: [nodes]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ name: kind, in: query, description: Filter by configured Node behavior, schema: { type: string, enum: [gateway, agent] } }]
      responses: { '200': { description: Nodes, content: { application/json: { schema: { $ref: '#/components/schemas/NodeList' } } } } }
    post:
      tags: [nodes]
      operationId: createNode
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/NodeInput' } } }
      responses: { '201': { description: Created }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /nodes/{nodeId}:
    parameters: [{ $ref: '#/components/parameters/NodeId' }]
    get:
      tags: [nodes]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses: { '200': { description: Node, content: { application/json: { schema: { $ref: '#/components/schemas/Node' } } } } }
    patch:
      tags: [nodes]
      operationId: updateNode
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/NodePatch' } } }
      responses: { '200': { description: Updated }, '409': { $ref: '#/components/responses/Conflict' } }
    delete:
      tags: [nodes]
      operationId: decommissionNode
      description: Decommission the Node identity instead of deleting its row. The Controller revokes admission, quarantines dependent Assignments, preserves Services and behavior specifications, and allows a later one-time replacement enrollment to reuse the same Node ID.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      responses: { '204': { description: Node decommissioned; the identity tombstone and dependent business resources are retained. }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /bootstrap/node/{asset}:
    get:
      tags: [operations]
      operationId: downloadNodeBootstrapAsset
      description: Download a Controller-provided Node installer or release descriptor using the one-time enrollment token. The asset endpoint is intentionally limited to the three generated bootstrap assets.
      security: [{ nodeEnrollmentAuth: [] }]
      parameters:
        - name: asset
          in: path
          required: true
          schema: { type: string, enum: [install-node.sh, install-node.ps1, release] }
      responses:
        '200': { description: Node installer script or release metadata, content: { text/plain: {}, text/x-shellscript: {}, application/json: {} } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Node bootstrap asset not found. }
        '503': { description: Node bootstrap asset is unavailable. }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    cookieAuth: { type: apiKey, in: cookie, name: af_session }
    nodeEnrollmentAuth: { type: apiKey, in: header, name: X-AsterFerry-Enrollment-Token, description: One-time token returned by the Dashboard node installation command generator. }
  parameters:
    NodeId: { name: nodeId, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 128 } }
    CSRF: { name: X-CSRF-Token, in: header, required: false, schema: { type: string, minLength: 32 } }
  responses:
    Unauthorized: { description: Authentication required }
    Conflict: { description: Revision or unique binding conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  schemas:
...
    NodeInput:
      type: object
      required: [id, name]
      properties: { id: { type: string }, name: { type: string }, labels: { type: object, additionalProperties: { type: string } }, enabled: { type: boolean } }
...
    NodePatch:
      type: object
      properties: { name: { type: string }, labels: { type: object, additionalProperties: { type: string } }, enabled: { type: boolean }, certificate_state: { type: string, enum: [pending, active, revoked, expired, decommissioned] }, certificate_serial: { type: string } }
    Node:
      allOf:
        - $ref: '#/components/schemas/NodeInput'
        - type: object
          properties: { spec_kind: { type: string, enum: [gateway, agent], description: Configured behavior projection; omitted until a spec is saved. }, certificate_state: { type: string, enum: [pending, active, revoked, expired, decommissioned] }, certificate_serial: { type: string }, revision: { type: integer, format: int64 }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } }
    NodeList: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Node' } } } }
...
    ErrorDetail: { type: object, properties: { code: { type: string }, message: { type: string }, path: { type: string }, retryable: { type: boolean } } }
    ErrorResponse: { type: object, properties: { error: { $ref: '#/components/schemas/ErrorDetail' } } }